Complete warehouse assistant application

+++ src/WarehouseAssistant.Api/Contracts/StockDtos.cs
@@ -3,4 +3,4 @@ namespace WarehouseAssistant.Api.Contracts;
 public sealed record StockRowResponse(
     int NomenclatureId,
     string NomenclatureName,
-    decimal Quantity);
+    int Quantity);
